restructure base routing
