+++ client/app/components/landingPage/landingPage.jsx
@@ -1,6 +1,9 @@
-import React from 'react';
+import React, { useState } from 'react';
+import LoginComponent from '../login/login';
 
 const LandingPage = () => {
+  const showModal = false;
+
   return (
     <div className="home-container">
       <div className="bg-img container-fluid"></div>
@@ -36,6 +39,7 @@ const LandingPage = () => {
           <p>Copyright &copy; 2020. All rights reserved</p>
         </div>
       </footer>
+      <LoginComponent showModal={showModal} />
     </div>
   );
 };
