+++ package.json
@@ -62,9 +62,7 @@
     "react-bootstrap": "^1.3.0",
     "react-dom": "^16.13.1",
     "react-redux": "^4.4.5",
-    "react-router": "^3.0.3",
     "react-router-dom": "^4.0.0",
-    "react-router-redux": "4.0.4",
     "redux": "3.5.2",
     "redux-devtools-extension": "^2.13.0",
     "redux-immutable-state-invariant": "^1.2.3",
